+++ src/NetzmachtContaoFormBundle.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Netzmacht\ContaoFormBundle;
 
 use Netzmacht\Contao\Toolkit\Bundle\DependencyInjection\Compiler\AddTaggedServicesAsArgumentPass;
+use Netzmacht\ContaoFormBundle\DependencyInjection\Compiler\TokenStoragePass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -34,5 +35,7 @@ final class NetzmachtContaoFormBundle extends Bundle
                 'netzmacht.contao_form.form_generator.mapper'
             )
         );
+
+        $container->addCompilerPass(new TokenStoragePass());
     }
 }